refactor: kshired 리뷰 반영

+++ src/main/kotlin/io/csbroker/apiserver/controller/v1/post/response/CommentResponseDto.kt
@@ -7,11 +7,11 @@ data class CommentResponseDto(
     val id: Long,
     val content: String,
     val username: String,
-    val likeCount: Int,
+    val likeCount: Long,
     val isLiked: Boolean,
     val createdAt: LocalDateTime,
 ) {
-    constructor(comment: Comment, likeCount: Int, isLiked: Boolean) : this(
+    constructor(comment: Comment, likeCount: Long, isLiked: Boolean) : this(
         id = comment.id,
         content = comment.content,
         username = comment.user.username,
